bugfix lbl flow; getHails(User|App); insertGuards
1.) Fixed bug in doForCollectionP: the flow check needs to be reversed

2.) Added support for extracting current user
and app (endoresed by corresponding principals).
TODO: make sure that Action is not instance of MonadState
since these functions use the headers.
An alternative implementaiton will set the session parameter
in HttpReq

3.) Added insertGuard which essentially does all the checks and taints
of insert, but does not perform an actual insert. This is useful for
cases where you (policy/service) want to do a save for an app that
normaly would not be able to do so while still tainting the app as its
performing an insert. In a sense, it's addressing the wguardP in saveP.